Issue #27748 move additional entity with experimental function
