+++ dev/io.openliberty.data.internal_fat/test-applications/DataTestApp/src/test/jakarta/data/web/DataTestServlet.java
@@ -21,7 +21,6 @@ import static org.junit.Assert.fail;
 import static test.jakarta.data.web.Assertions.assertIterableEquals;
 
 import java.time.Duration;
-import java.time.OffsetDateTime;
 import java.time.Year;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -129,9 +128,6 @@ public class DataTestServlet extends FATServlet {
 
     @Inject
     Receipts receipts;
-
-    @Inject
-    Shipments shipments;
 
     @Inject
     Things things;
@@ -3474,88 +3470,6 @@ public class DataTestServlet extends FATServlet {
     public void testParameterAnnotationTakesPrecedenceOverMethodPrefix() {
         Prime nine = primes.findByBinary("10011").orElseThrow();
         assertEquals(19L, nine.numberId);
-    }
-
-    /**
-     * Invoke methods that are annotated with the Select, Where, Update, and Delete annotations.
-     */
-    @Test
-    public void testPartialQueryAnnotations() {
-        Shipment s1 = new Shipment();
-        s1.setDestination("200 1st Ave SW, Rochester, MN 55902");
-        s1.setLocation("44.027354, -92.468482");
-        s1.setId(1);
-        s1.setOrderedAt(OffsetDateTime.now().minusMinutes(45));
-        s1.setStatus("IN_TRANSIT");
-        shipments.save(s1);
-
-        Shipment s2 = new Shipment();
-        s2.setDestination("201 4th St SE, Rochester, MN 55904");
-        s2.setLocation("2800 37th St NW, Rochester, MN 55901");
-        s2.setId(2);
-        s2.setOrderedAt(OffsetDateTime.now().minusMinutes(20));
-        s2.setStatus("READY_FOR_PICKUP");
-        shipments.save(s2);
-
-        Shipment s3 = new Shipment();
-        s3.setDestination("151 4th St SE, Rochester, MN 55904");
-        s3.setLocation("44.057840, -92.496301");
-        s3.setId(3);
-        s3.setOrderedAt(OffsetDateTime.now().minusMinutes(13));
-        s3.setStatus("IN_TRANSIT");
-        shipments.save(s3);
-
-        Shipment s4 = new Shipment();
-        s4.setDestination("151 4th St SE, Rochester, MN 55904");
-        s4.setLocation("2800 37th St NW, Rochester, MN 55901 ");
-        s4.setId(4);
-        s4.setOrderedAt(OffsetDateTime.now().minusMinutes(4));
-        s4.setStatus("READY_FOR_PICKUP");
-        shipments.save(s4);
-
-        Shipment s5 = new Shipment();
-        s5.setDestination("201 4th St SE, Rochester, MN 55904");
-        s5.setLocation("2800 37th St NW, Rochester, MN 55901");
-        s5.setId(5);
-        s5.setOrderedAt(OffsetDateTime.now().minusSeconds(50));
-        s5.setStatus("PREPARING");
-        shipments.save(s5);
-
-        assertEquals(true, shipments.dispatch(2, "READY_FOR_PICKUP",
-                                              "IN_TRANSIT", "44.036217, -92.488040", OffsetDateTime.now()));
-        assertEquals("IN_TRANSIT", shipments.getStatus(2));
-
-        // @OrderBy "destination"
-        assertIterableEquals(List.of("151 4th St SE, Rochester, MN 55904",
-                                     "200 1st Ave SW, Rochester, MN 55902",
-                                     "201 4th St SE, Rochester, MN 55904"),
-                             shipments.find("IN_TRANSIT")
-                                             .map(o -> o.getDestination())
-                                             .collect(Collectors.toList()));
-
-        // @OrderBy "status", then "orderedAt" descending
-        assertIterableEquals(List.of(3L, 2L, 1L, 5L, 4L),
-                             Stream.of(shipments.getAll()).map(o -> o.getId()).collect(Collectors.toList()));
-
-        Shipment s = shipments.find(3);
-        String previousLocation = s.getLocation();
-
-        assertEquals(true, shipments.updateLocation(3, previousLocation, "44.029468, -92.483191"));
-        assertEquals(false, shipments.updateLocation(3, previousLocation, "44.029406, -92.489553"));
-
-        s = shipments.find(3);
-        assertEquals("44.029468, -92.483191", s.getLocation());
-
-        assertEquals(true, shipments.cancel(4, Set.of("PREPARING", "READY_FOR_PICKUP"),
-                                            "CANCELED", OffsetDateTime.now()));
-        assertEquals(true, shipments.cancel(5, Set.of("PREPARING", "READY_FOR_PICKUP"),
-                                            "CANCELED", OffsetDateTime.now()));
-        assertEquals(false, shipments.cancel(10, Set.of("PREPARING", "READY_FOR_PICKUP"),
-                                             "CANCELED", OffsetDateTime.now()));
-
-        assertEquals(2, shipments.statusBasedRemoval("CANCELED"));
-
-        assertEquals(3, shipments.removeEverything());
     }
 
     /**
